+++ src/bl64-pkg-env.bash
@@ -4,7 +4,7 @@
 
 # shellcheck disable=SC2034
 {
-  declare BL64_PKG_VERSION='6.5.5'
+  declare BL64_PKG_VERSION='6.6.0'
 
   declare BL64_PKG_MODULE='0'