+++ src/bl64-ui-env.bash
@@ -4,9 +4,10 @@
 
 # shellcheck disable=SC2034
 {
-  declare BL64_UI_VERSION='3.2.0'
+  declare BL64_UI_VERSION='3.3.0'
 
   declare BL64_UI_MODULE='0'
 
   declare BL64_UI_CONFIRMATION_TIMEOUT='60'
+  declare BL64_UI_CONFIRMATION_SKIP="$BL64_VAR_NO"
 }